Codecov pull request checks: use non-deprecated v2 version of action
As documented on https://github.com/codecov/codecov-action, v1 of
Codecov's GitHub action will cease to function on 2022-02-01. Given that
we do not seem to use any advanced flags, the migration should not
require changes.

It does, however, permit one change: as CBMC is a public project, a
token isn't even necessary for the upload.