unwindset: Also warn about loops in functions without body
a477e19 demoted some failures to warnings. This commit does so for
another case: if no body for the function is available, issue a warning
rather than failing an exception.

This situation will typically arise when using an unwindset
specification originally meant for CBMC with goto-instrument, where
library functions might not yet be present.