Simplification of nested pointer arithmetic: do not assume same type
Pointer arithmetic is permitted with both signed and unsigned numeric
operands. Nested expressions could possibly mix this when the front-end
does not strictly use the same type for pointer arithmetic as for index
types. Observed in https://github.com/model-checking/rmc/issues/708.

Introducing type casts (as needed) exposed a further issue in this code:
it previously did not make sure that it was really the numeric operands
that ended up being grouped together. This became apparent on test
Pointer_Arithmetic15, which specifically constructs such expressions.