Value sets must not return an empty set for nondet symbols
Even when the expression type is not immediately a pointer, some
component of the type may be a pointer. As such, it could legitimately
appear in pointer dereferencing, and we should at least add "unknown" to
the value set.

This is a follow-up fix to c14e22a41: the regression test included
yields a spurious counterexample in versions between c14e22a41 and this
bugfix commit. This commit also reverts the (spurious) changes to the
double_deref/* regression tests.