Add annotated_pointer_constant_exprt
The back-end previously created constant_exprt with an operand when
instantiating pointer-typed expressions with their solver-computed
model. This violates the contract of nullary_exprt, which a
constant_exprt should (!) be, and resulted in special-case treatment in
several places.

The annotated_pointer_constant_exprt formalises this concept of a
constant (as derived from the solver-computed model) with the additional
annotation of the original pointer that was instantiated.

Fixes: #6691