Function pointer restrictions: refactor pointer-name parsing
No changes in behaviour: this commit just moves the pointer-name parsing
introduced in cc506551892fecb to a function of its own. This is to
prepare for a fix of #6609: the same name parsing also needs to be added
to function-pointer-restriction parsing from JSON files.

To enable this code move, the exception class is moved to the global
scope.