C front-end: refactor type checking __builtin_*_overflow
Move the implementation to a method of its own. This avoids repeatedly
testing the name to derive the underlying arithmetic operation. This
change is to support upcoming extensions to handle further built-ins via
this code path. No changes in behaviour intended with just this commit,
and additional testing should ensure that this is also true for error
handling.

The function refactored from remains too large, and cpplint rightly
complains about that, but this issue cannot be fixed in a single commit.
Hence, adding an override.