Switch to clang-format-10 to support more recent platforms
clang-format-7 (or older) might not even be available anymore on some
systems, and will certainly not be the default. clang-format-10 has
improved support for initializer lists, resulting in different
formatting of these:
```
foo bar{value_1,
        value_2};
```
now is formatted as
```
foo bar{
  value_1,
  value_2};
```
which is more consistent with how the same would have been formatted
had parentheses been used.

clang-format-10 is even available on Ubuntu 18.04, which is the most
dated system our CI builds are still running on.

Fixes: #6457