Add indexed identifier support to smt function `factoryt`
This is implemented using sfinae to detect if the supplied `functiont`
has an `indices` member function or not, so that it is unneccessary to
add an `indices` member function which returns the empty collections of
indicies for the cases where `function` needs a simple (non indexed)
identifier.

`type_traits.h` is in `solvers/smt2_incremental/` because this is the
only part of the codebase where it is currently used and keeping it here
means the PR with this commit should need fewer code reviewer approvals.