Do pretty printing for debug logging only when vebosity is high
This avoids wasting execution time in `irept::pretty` in the case where
verbosity is set to the default level. Before this performance fix, the
pretty printed output was generated and then discarded when the
verbosity was set to the default level.

Comparing this change against the `develop` branch, using a benchmark
has been shown to reduce the runtime from approximately 12 seconds to
approximately 3 seconds for that particular example.