Make has_symbol configurable and deprecate the existing has_symbol
The existing code did not permit configuring what kinds of symbols are
to be found. It had a single user within the code base, and even that
one was not a good use of it for it constructed a single-element set.

The new implementation does not yet make use of potential efficiency
advantages of "has a symbol" over "find the existing symbols," but such
improvements will be done based on further cleanup of find_symbols code.