CONTRACTS: Error-out on unknown functions (replace/enforce).
We dropped the pattern of returning boolean error flags and accumulating
errors in `replace_calls` and `enforce_contracts`.
We are now checking that all functions exist upfront and
throwing exceptions from `util/exception_utils.h` as soon as the first
unknown function is found.
Subsequent checks of the same conditions are now INVARIANTS.