rename_symbolt: renaming in return values should affect the result
If users want to check whether any renaming took place, they need to be
informed in case that renaming took place in the type of the return
value of a code_typet.

Only function_name_manglert currently checks this result value, but none
of its current use would really result in renaming within a return
value.